backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
pandas==2.3.2
openpyxl==3.1.5
python-multipart==0.0.20
supabase==2.15.0
